Squash additional migrations to prior ones
Since #4295 enforces complete DB resync on some nodes, this PR also rebases the migrations by squashing the additional migrations into the previous ones. This simplifies the migration code.

Closes #6180
