+++ db/migration/src/lib.rs
@@ -12,10 +12,6 @@ mod m20240226_000009_peers_create_peer_store;
 mod m20240301_000010_tickets_create_ticket;
 mod m20240301_000011_tickets_create_ticket_stats;
 mod m20240301_000012_tickets_create_outgoing_ticket_index;
-mod m20240301_000013_initial_seed_tickets;
-mod m20240301_000014_create_ticket_stats_with_channel_id;
-mod m20240326_000015_recreate_ticket_stats;
-mod m20240418_000016_add_rolling_tx_hash;
 
 #[derive(PartialEq)]
 pub enum BackendType {
@@ -41,16 +37,12 @@ impl MigratorTrait for Migrator {
             Box::new(m20240226_000008_node_create_settings::Migration),
             Box::new(m20240226_000009_peers_create_peer_store::Migration),
             Box::new(m20240301_000010_tickets_create_ticket::Migration(BackendType::Postgres)),
-            Box::new(m20240301_000011_tickets_create_ticket_stats::Migration),
+            Box::new(m20240301_000011_tickets_create_ticket_stats::Migration(
+                BackendType::Postgres,
+            )),
             Box::new(m20240301_000012_tickets_create_outgoing_ticket_index::Migration(
                 BackendType::Postgres,
             )),
-            Box::new(m20240301_000013_initial_seed_tickets::Migration),
-            Box::new(m20240301_000014_create_ticket_stats_with_channel_id::Migration(
-                BackendType::Postgres,
-            )),
-            Box::new(m20240326_000015_recreate_ticket_stats::Migration(BackendType::Postgres)),
-            Box::new(m20240418_000016_add_rolling_tx_hash::Migration),
         ]
     }
 }
@@ -74,7 +66,6 @@ impl MigratorTrait for MigratorIndex {
             Box::new(m20240226_000006_index_create_network_eligibility::Migration),
             Box::new(m20240226_000008_node_create_settings::Migration),
             Box::new(m20240226_000007_index_initial_seed::Migration),
-            Box::new(m20240418_000016_add_rolling_tx_hash::Migration),
         ]
     }
 }
@@ -95,15 +86,12 @@ impl MigratorTrait for MigratorTickets {
     fn migrations() -> Vec<Box<dyn MigrationTrait>> {
         vec![
             Box::new(m20240301_000010_tickets_create_ticket::Migration(BackendType::SQLite)),
-            Box::new(m20240301_000011_tickets_create_ticket_stats::Migration),
+            Box::new(m20240301_000011_tickets_create_ticket_stats::Migration(
+                BackendType::SQLite,
+            )),
             Box::new(m20240301_000012_tickets_create_outgoing_ticket_index::Migration(
                 BackendType::SQLite,
             )),
-            Box::new(m20240301_000013_initial_seed_tickets::Migration),
-            Box::new(m20240301_000014_create_ticket_stats_with_channel_id::Migration(
-                BackendType::SQLite,
-            )),
-            Box::new(m20240326_000015_recreate_ticket_stats::Migration(BackendType::SQLite)),
         ]
     }
 }
