+++ db/migration/src/m20240301_000011_tickets_create_ticket_stats.rs
@@ -1,8 +1,8 @@
-use crate::sea_orm::prelude::ChronoDateTimeUtc;
+use crate::BackendType;
 use sea_orm_migration::prelude::*;
 
 #[derive(DeriveMigrationName)]
-pub struct Migration;
+pub struct Migration(pub BackendType);
 
 #[async_trait::async_trait]
 impl MigrationTrait for Migration {
@@ -20,19 +20,14 @@ impl MigrationTrait for Migration {
                             .primary_key(),
                     )
                     .col(
-                        ColumnDef::new(TicketStatistics::LastUpdated)
-                            .timestamp()
+                        ColumnDef::new(TicketStatistics::ChannelId)
+                            .string_len(64)
                             .not_null()
-                            .default(ChronoDateTimeUtc::UNIX_EPOCH),
+                            .unique_key(),
                     )
+                    .col(ColumnDef::new(TicketStatistics::LastUpdated).timestamp().null())
                     .col(
-                        ColumnDef::new(TicketStatistics::LosingTickets)
-                            .not_null()
-                            .integer()
-                            .default(0),
-                    )
-                    .col(
-                        ColumnDef::new(TicketStatistics::RedeemedTickets)
+                        ColumnDef::new(TicketStatistics::WinningTickets)
                             .not_null()
                             .integer()
                             .default(0),
@@ -44,22 +39,10 @@ impl MigrationTrait for Migration {
                             .default(vec![0u8; 12]),
                     )
                     .col(
-                        ColumnDef::new(TicketStatistics::NeglectedTickets)
-                            .not_null()
-                            .integer()
-                            .default(0),
-                    )
-                    .col(
                         ColumnDef::new(TicketStatistics::NeglectedValue)
                             .binary_len(12)
                             .not_null()
                             .default(vec![0u8; 12]),
-                    )
-                    .col(
-                        ColumnDef::new(TicketStatistics::RejectedTickets)
-                            .not_null()
-                            .integer()
-                            .default(0),
                     )
                     .col(
                         ColumnDef::new(TicketStatistics::RejectedValue)
@@ -69,10 +52,33 @@ impl MigrationTrait for Migration {
                     )
                     .to_owned(),
             )
-            .await
+            .await?;
+
+        let conn = manager.get_connection();
+
+        conn.execute_unprepared(
+            r#"
+            CREATE TRIGGER IF NOT EXISTS trig_ticket_stats_update_timestamp
+                AFTER UPDATE
+                ON ticket_statistics
+                FOR EACH ROW
+                WHEN OLD.last_updated IS NULL OR NEW.last_updated < OLD.last_updated --- avoid infinite loop
+            BEGIN
+                UPDATE ticket_statistics SET last_updated=CURRENT_TIMESTAMP WHERE id=OLD.id;
+            END;
+        "#,
+        )
+        .await?;
+
+        Ok(())
     }
 
     async fn down(&self, manager: &SchemaManager) -> Result<(), DbErr> {
+        let conn = manager.get_connection();
+
+        conn.execute_unprepared("DROP TRIGGER trig_ticket_stats_update_timestamp;")
+            .await?;
+
         manager
             .drop_table(Table::drop().table(TicketStatistics::Table).to_owned())
             .await
@@ -83,12 +89,10 @@ impl MigrationTrait for Migration {
 enum TicketStatistics {
     Table,
     Id,
+    ChannelId,
     LastUpdated,
-    LosingTickets,
-    RedeemedTickets,
+    WinningTickets,
     RedeemedValue,
-    NeglectedTickets,
     NeglectedValue,
-    RejectedTickets,
     RejectedValue,
 }
